feat: circular-dependency check & isolatedModules in tsconfig
