+++ package.json
@@ -48,6 +48,7 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.6.1",
+    "@types/circular-dependency-plugin": "^5.0.8",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.21",
     "@types/react": "^18.0.17",
@@ -59,6 +60,7 @@
     "@typescript-eslint/eslint-plugin": "^8.16.0",
     "@typescript-eslint/parser": "^8.16.0",
     "babel-loader": "^8.2.3",
+    "circular-dependency-plugin": "^5.2.2",
     "clsx": "^2.1.1",
     "concurrently": "^9.1.0",
     "copy-webpack-plugin": "^12.0.2",
