+++ src/app/providers/StoreProvider/config/store.ts
@@ -42,5 +42,3 @@ export function createReduxStore(
 
   return store;
 }
-
-export type AppDispatch = ReturnType<typeof createReduxStore>['dispatch'];
