+++ src/entities/Article/index.ts
@@ -1,9 +1,9 @@
 export { Article } from './ui/Article/Article';
 export type { Article as ArticleModel } from './model/types/article';
+export type { ArticleTextBlock } from './model/types/article';
 export {
-  ArticleSortField, ArticleType, ArticleBlockType, ArticleTextBlock,
-} from './model/types/article';
+  ArticleSortField, ArticleType, ArticleBlockType, ArticleListView,
+} from './model/consts/consts';
 export type { ArticleDetailsModel } from './model/types/articleDetailsModel';
-export { ArticleListView } from './model/types/article';
 export { ArticleList } from './ui/ArticleList/ArticleList';
 export { getArticleData } from './model/selectors/article';
