+++ src/entities/Article/model/consts/consts.ts
@@ -1,0 +1,24 @@
+export enum ArticleType {
+  ALL = 'ALL',
+  IT = 'IT',
+  SCIENCE = 'SCIENCE',
+  ECONOMICS = 'ECONOMICS'
+}
+
+export enum ArticleBlockType {
+  CODE = 'CODE',
+  IMAGE = 'IMAGE',
+  TEXT = 'TEXT',
+}
+
+export enum ArticleListView {
+  LIST = 'list',
+  TILE = 'tile',
+  LINE = 'line',
+}
+
+export enum ArticleSortField {
+  VIEWS = 'views',
+  TITLE = 'title',
+  CREATED = 'createdAt',
+}
