+++ src/entities/Article/model/types/article.ts
@@ -1,29 +1,5 @@
 import { User } from 'entities/User';
-
-export enum ArticleType {
-  ALL = 'ALL',
-  IT = 'IT',
-  SCIENCE = 'SCIENCE',
-  ECONOMICS = 'ECONOMICS'
-}
-
-export enum ArticleBlockType {
-  CODE = 'CODE',
-  IMAGE = 'IMAGE',
-  TEXT = 'TEXT',
-}
-
-export enum ArticleListView {
-  LIST = 'list',
-  TILE = 'tile',
-  LINE = 'line',
-}
-
-export enum ArticleSortField {
-  VIEWS = 'views',
-  TITLE = 'title',
-  CREATED = 'createdAt',
-}
+import { ArticleType, ArticleBlockType } from '../consts/consts';
 
 export interface ArticleBlockBase {
   id: string;
