+++ src/entities/Article/ui/Article/Article.tsx
@@ -14,8 +14,9 @@ import { TitleSize } from 'shared/ui/Title/Title';
 import { useMediaQuery } from '@mantine/hooks';
 import { getArticleById } from '../../model/services/getArticleById/getArticleById';
 import { articleReducer } from '../../model/slices/articleSlice';
-import { getArticleData, getArticleError, getArticleIsLoading } from '../../model/selectors/article';
-import { ArticleBlock, ArticleBlockType } from '../../model/types/article';
+import { getArticleData, getArticleError } from '../../model/selectors/article';
+import { ArticleBlock } from '../../model/types/article';
+import { ArticleBlockType } from '../../model/consts/consts';
 import { ArticleCodeBlockComponent } from '../ArticleCodeBlockComponent/ArticleCodeBlockComponent';
 import { ArticleImageBlockComponent } from '../ArticleImageBlockComponent/ArticleImageBlockComponent';
 import { ArticleTextBlockComponent } from '../ArticleTextBlockComponent/ArticleTextBlockComponent';
