+++ src/entities/Article/ui/ArticleListItem/ArticleListItemSkeleton.tsx
@@ -1,7 +1,7 @@
 import cx from 'clsx';
 import { memo } from 'react';
 import { Card, Skeleton } from 'shared/ui';
-import { ArticleListView } from '../../model/types/article';
+import { ArticleListView } from '../../model/consts/consts';
 
 import classes from './ArticleListItem.module.css';
