+++ src/entities/Profile/index.ts
@@ -1,2 +1,2 @@
-export { Profile } from './model/types/profileType';
+export type { Profile } from './model/types/profileType';
 export { ProfileCard } from './ui/ProfileCard/ProfileCard';
