+++ src/entities/User/index.ts
@@ -1,4 +1,5 @@
 export { userReducer, userActions } from './model/slice/userSlice';
-export { UserModel, User, UserRole } from './model/types/userModel';
+export type { UserModel, User } from './model/types/userModel';
+export { UserRole } from './model/consts/consts';
 export { getUserAuthData, getUserMounted } from './model/selectors/user';
 export { isUserAdmin, isUserManager, getUserRoles } from './model/selectors/roleSelectors';
