+++ src/features/AddCommentForm/index.ts
@@ -1,2 +1,2 @@
-export { AddCommentFormModel } from './model/types/addCommentForm';
+export type { AddCommentFormModel } from './model/types/addCommentForm';
 export { AddCommentFormAsync as AddCommentForm } from './ui/AddCommentForm/AddCommentFormAsync';
