+++ src/features/AuthUserName/index.ts
@@ -1,3 +1,2 @@
 export { LoginModal } from './ui/LoginModal/LoginModal';
-// export { loginReducer } from './model/slice/loginSlice';
-export { LoginModel } from './model/types/loginModel';
+export type { LoginModel } from './model/types/loginModel';
