+++ src/features/EditableProfileCard/index.ts
@@ -1,3 +1,3 @@
 export { EditableProfileCard } from './ui/EditableProfileCard/EditableProfileCard';
-export { ProfileModel } from './model/types/EditableProfileCardModel';
+export type { ProfileModel } from './model/types/EditableProfileCardModel';
 export { EditableProfileCardHeader } from './ui/EditableProfileCardHeader/EditableProfileCardHeader';
