+++ src/features/EditableProfileCard/model/consts/consts.ts
@@ -1,0 +1,7 @@
+export enum ValidateProfileError {
+  INCORRECT_USER_DATA = 'incorrect_user_data',
+  INCORRECT_AGE = 'incorrect_age',
+  INCORRECT_CITY = 'incorrect_city',
+  NO_DATA = 'not_data',
+  SERVER_ERROR = 'server error',
+}
