+++ src/features/EditableProfileCard/model/types/EditableProfileCardModel.ts
@@ -1,12 +1,5 @@
 import { Profile } from 'entities/Profile';
-
-export enum ValidateProfileError {
-  INCORRECT_USER_DATA = 'incorrect_user_data',
-  INCORRECT_AGE = 'incorrect_age',
-  INCORRECT_CITY = 'incorrect_city',
-  NO_DATA = 'not_data',
-  SERVER_ERROR = 'server error',
-}
+import { ValidateProfileError } from '../consts/consts';
 
 export interface ProfileModel {
   data?: Profile;
