+++ src/features/ScrollSave/index.ts
@@ -1,4 +1,4 @@
-export { ScrollSaveModel } from './model/types/scrollSaveModel';
+export type { ScrollSaveModel } from './model/types/scrollSaveModel';
 
 export { getScrollSaveByPath } from './model/selectors/scrollSave';
 export { scrollSaveReducer, setScrollPosition } from './model/slice/scrollSaveSlice';
