+++ src/shared/lib/index.ts
@@ -1,6 +1,4 @@
 export { classNames } from './classNames/classNames';
-export {
-  DynamicModuleLoader,
-  ReducersList,
-} from './components/DynamicModuleLoader/DynamicModuleLoader';
+export type { ReducersList } from './components/DynamicModuleLoader/DynamicModuleLoader';
+export { DynamicModuleLoader } from './components/DynamicModuleLoader/DynamicModuleLoader';
 export { addQueryParams } from './url/addQueryParams/addQueryParams';
