+++ tsconfig.json
@@ -14,6 +14,7 @@
     "paths": {
       "*": ["./src/*"]
     },
+    "isolatedModules": true,
   },
   "ts-node": {
     "compilerOptions": {
